Refactor multiple-choice, option button modal opening at button
